address pr review comments
- ci.yaml: revert TEMP workflow_dispatch widening
- ui.rs: fold ui_impl back into ui.rs (single file, no #[path] shim)
- api.rs: collapse cfg(unix)/cfg(windows) serve_session pair into a single
  non-cfg-gated function. Push the bind/cleanup logic into transport_unix.rs
  and transport_windows.rs, both exposing bind_session_transport() ->
  (Listener, SessionTransportCleanup)
- transport_windows.rs: rename sa -> security_attributes
- win_security.rs: NULL-guard CloseHandle in TokenHandle::drop;
  rename sa -> security_attributes; switch *mut SECURITY_ATTRIBUTES cast
  end to *mut _; drop the 1)/2)/3)/4) numbering; remove redundant
  comments; add safety comments on the Vec<u8> 1-byte-aligned casts;
  link to the ACL winapi docs; rephrase the ACE comment
- session-monitor-client/lib.rs: use winapi's named ERROR_PIPE_BUSY
  constant instead of the magic 231
- xtask/monitor.rs: link to MS docs in the PATHEXT comment
