+++ mirrord/session-monitor-client/Cargo.toml
@@ -31,6 +31,9 @@ serde_json.workspace = true
 tokio = { workspace = true, features = ["net", "io-util", "rt", "time"] }
 tracing.workspace = true
 
+[target.'cfg(windows)'.dependencies]
+winapi = { workspace = true }
+
 [lib]
 test = false
 doctest = false
